[router-server] Split createPages into its own file
