+++ packages/@expo/router-server/src/rsc/router/noopRouter.ts
@@ -1,6 +1,6 @@
 import type { EntriesDev } from '../server';
 
-import { createPages } from './expo-definedRouter';
+import { createPages } from './createPages';
 
 // Used in client-only mode to skip route loading in react-server bundles.
 export default (_getRouteOptions?: unknown): EntriesDev => ({
